move hurst and stats modules
