+++ setup.cfg
@@ -21,7 +21,6 @@ pacakges=find:
 install_requires = 
     numpy
     pandas
-    scipy
 
 [options.packages.find]
 where=src
