+++ src/technical_analysis/hurst.py
@@ -3,7 +3,7 @@ from typing import Tuple, Union
 import numpy as np
 import pandas as pd
 
-from technical_analysis.stats.stats import linear_regression
+from technical_analysis.stats import linear_regression
 
 
 def rescaled_range(arr: np.ndarray):
